gcode: New lexer, parser and interpreter
The lexer is an incremental single-pass implementation in C.  It uses gperf for
efficient lookup of tokens and focuses on performance.  It's a bit more complex
than I'd prefer but we mitigate this with comprehensive tests (see below).

The parser is also incremental, implemented using a Bison push parser.  It's
output is an AST also defined in this commit.  It supports current Klipper
G-Code syntax with the addition of embedded expressions.

The "interpreter" increments through statements and evaluates expressions, but
is not a full interpreter in the traditional sense.  Rather, it outputs raw
G-Code in the form of (command, parameter list) tuples appropriate for
ingestion by Klipper's Python G-Code interpreter.

This commit also includes build infrastructure.  The makefile in
gcode_support/ builds a command line utility and uses this utility to test
above.

Tests work by interpreting test.input.gcode and comparing the result against
test.output.gcode.  The input file attempts to comprehensive cover syntax and
expression evaluation.

This commit only implements adds these modules.  It does not wire the new
modules into Klippy.

Signed-off-by: Greg Lauckhart <greg@lauckhart.com>